requirements.txt
Flask==2.2.2
opencv-python==4.7.0.68
numpy==1.23.4
Pillow==9.3.0
torch==2.0.1
transformers==4.30.0